libm: Fix SAFE_*_SHIFT with negative shift amounts
These were comparing integers with size_t values, which could result
in the int being cast to unsigned and the compare being wrong for
negative values. Instead, cast the size to int so that the compare is
done with signed values.

+++ newlib/libm/common/fdlibm.h
@@ -372,10 +372,10 @@ do {								\
    of a shift is exactly equal to the size of the shifted operand.  */
 
 #define SAFE_LEFT_SHIFT(op,amt)					\
-  (((amt) < 8 * sizeof(op)) ? ((op) << (amt)) : 0)
+  (((amt) < (int) (8 * sizeof(op))) ? ((op) << (amt)) : 0)
 
 #define SAFE_RIGHT_SHIFT(op,amt)				\
-  (((amt) < 8 * sizeof(op)) ? ((op) >> (amt)) : 0)
+  (((amt) < (int) (8 * sizeof(op))) ? ((op) >> (amt)) : 0)
 
 #ifdef  _COMPLEX_H
